test(musefs-format): dedup FLAC/WAV fixtures and layout helpers, rename FLAC tests (closes #451)
Rename the un-prefixed FLAC integration tests to the `<format>_<concern>`
convention used by every other format (locate.rs -> flac_locate.rs, etc.).

Hoist the duplicated helpers into tests/common/mod.rs:
- flac_fixture(fill, len) replaces the near-identical fixture() in the two
  FLAC synthesize tests.
- fmt_pcm_16bit_mono() and build_wav() replace the WAV RIFF builders copied
  across wav_locate, wav_read_tags, wav_synthesize, and proptest_wav.

Replace the five reimplemented layout-flattening helpers (assemble/materialize/
materialize_wav) with the existing common::resolve_layout, which already covers
every segment kind, in wav_synthesize, mp3_synthesize, mp4_oracle, proptest_mp4,
and proptest_wav.
