+++ musefs-format/tests/wav_locate.rs
@@ -1,36 +1,8 @@
+mod common;
+
+use common::{build_wav, fmt_pcm_16bit_mono};
 use musefs_format::FormatError;
 use musefs_format::wav::{locate_audio, read_structure};
-
-/// A 16-byte PCM `fmt ` payload: mono, 44.1 kHz, 16-bit.
-pub fn fmt_pcm_16bit_mono() -> Vec<u8> {
-    let mut f = Vec::new();
-    f.extend_from_slice(&1u16.to_le_bytes()); // wFormatTag = PCM
-    f.extend_from_slice(&1u16.to_le_bytes()); // channels
-    f.extend_from_slice(&44_100u32.to_le_bytes()); // sample rate
-    f.extend_from_slice(&88_200u32.to_le_bytes()); // byte rate
-    f.extend_from_slice(&2u16.to_le_bytes()); // block align
-    f.extend_from_slice(&16u16.to_le_bytes()); // bits per sample
-    f
-}
-
-/// Build a minimal valid `RIFF/WAVE` file from a list of `(fourcc, payload)` chunks.
-pub fn build_wav(chunks: &[(&[u8; 4], Vec<u8>)]) -> Vec<u8> {
-    let mut body = Vec::new();
-    for (id, payload) in chunks {
-        body.extend_from_slice(*id);
-        body.extend_from_slice(&u32::try_from(payload.len()).unwrap().to_le_bytes());
-        body.extend_from_slice(payload);
-        if payload.len() % 2 == 1 {
-            body.push(0x00);
-        }
-    }
-    let mut out = Vec::new();
-    out.extend_from_slice(b"RIFF");
-    out.extend_from_slice(&u32::try_from(body.len() + 4).unwrap().to_le_bytes());
-    out.extend_from_slice(b"WAVE");
-    out.extend_from_slice(&body);
-    out
-}
 
 #[test]
 fn locate_finds_data_bounds() {
